test: Test backends
