+++ tests/backends/cassettes/test_slack/TestSlackBackend.test_send_message.yaml
@@ -1,0 +1,19 @@
+interactions:
+- request:
+    body: '{"channel": "<REDACTED>", "text": "Hello, World!"}'
+    headers: {}
+    method: POST
+    uri: https://www.slack.com/api/chat.postMessage
+  response:
+    body:
+      string: '{"ok": true, "channel": "<REDACTED>", "ts": "1703857095.689759", "message":
+        {"bot_id": "<REDACTED>", "type": "message", "text": "Hello, World!", "user":
+        "<REDACTED>", "ts": "1703857095.689759", "app_id": "<REDACTED>", "blocks":
+        [{"type": "rich_text", "block_id": "JVS4", "elements": [{"type": "rich_text_section",
+        "elements": [{"type": "text", "text": "Hello, World!"}]}]}], "team": "<REDACTED>",
+        "bot_profile": {}}}'
+    headers: {}
+    status:
+      code: 200
+      message: OK
+version: 1
