+++ tests/backends/cassettes/test_slack/TestSlackRedirectBackend.test_send_message.yaml
@@ -1,22 +1,22 @@
 interactions:
 - request:
     body: '{"channel": "<REDACTED>", "text": "Hello, World!", "attachments": [{"color":
-      "#eb4034", "text": ":warning:  This message was originally sent to channel *whatever-channel*
+      "#eb4034", "text": ":warning:  This message was originally sent to channel *whatever-this-channel*
       but redirected here."}]}'
     headers: {}
     method: POST
     uri: https://www.slack.com/api/chat.postMessage
   response:
     body:
-      string: '{"ok": true, "channel": "<REDACTED>", "ts": "1703857089.884559", "message":
+      string: '{"ok": true, "channel": "<REDACTED>", "ts": "1703857096.171159", "message":
         {"bot_id": "<REDACTED>", "type": "message", "text": "Hello, World!", "user":
-        "<REDACTED>", "ts": "1703857089.884559", "app_id": "<REDACTED>", "blocks":
-        [{"type": "rich_text", "block_id": "FbU8", "elements": [{"type": "rich_text_section",
+        "<REDACTED>", "ts": "1703857096.171159", "app_id": "<REDACTED>", "blocks":
+        [{"type": "rich_text", "block_id": "Qn2", "elements": [{"type": "rich_text_section",
         "elements": [{"type": "text", "text": "Hello, World!"}]}]}], "team": "<REDACTED>",
         "bot_profile": {}, "attachments": [{"id": 1, "color": "eb4034", "fallback":
-        ":warning:  This message was originally sent to channel *whatever-channel*
+        ":warning:  This message was originally sent to channel *whatever-this-channel*
         but redirected here.", "text": ":warning:  This message was originally sent
-        to channel *whatever-channel* but redirected here."}]}}'
+        to channel *whatever-this-channel* but redirected here."}]}}'
     headers: {}
     status:
       code: 200
